+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+
+- New `--latest` flag to `sedge generate` subcommands to use the latest version of the image in the clients.
+
 ### Fixed
 
 - Remove TTD of the params, test cases, command flags and geth patch for custom testnets.